more updates
grid and parent image are lined up and sit on background just fine when viewport isn't too wide. Need to resave the image of the house as it looks wonky. Will do that in the morning, and work on the nav system so I can work on the "inside" illustrations